+++ src/components/UserDetailSections/UserAffiliations/UserAffiliations.test.js
@@ -1,16 +1,33 @@
-import { screen } from '@testing-library/react';
-
+import {
+  screen,
+  waitForElementToBeRemoved
+} from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import {
+  QueryClient,
+  QueryClientProvider,
+} from 'react-query';
 import renderWithRouter from 'helpers/renderWithRouter';
 
 import affiliations from '../../../../test/jest/fixtures/affiliations';
-import { useUserAffiliations } from '../../../hooks';
+import {
+  useUserAffiliations,
+  useUserAffiliationsMutation
+} from '../../../hooks';
 import UserAffiliations from './UserAffiliations';
+import { getResponseErrors } from './util';
+
+const queryClient = new QueryClient();
 
 jest.unmock('@folio/stripes/components');
 jest.mock('../../../hooks', () => ({
   ...jest.requireActual('../../../hooks'),
-  useUserAffiliations: jest.fn(() => ({ affiliations: [], totalRecords: 0 })),
-  useUserAffiliationsMutation: jest.fn(() => ({ handleAssignment: jest.fn(), isLoading: false })),
+  useUserAffiliations: jest.fn(),
+  useUserAffiliationsMutation: jest.fn(),
+}));
+
+jest.mock('./util', () => ({
+  getResponseErrors: jest.fn(() => []),
 }));
 
 const defaultProps = {
@@ -18,20 +35,24 @@ const defaultProps = {
   expanded: true,
   onToggle: jest.fn(),
   userId: 'userId',
+  userName: 'mobius',
 };
 
 const renderPatronBlock = (props = {}) => renderWithRouter(
-  <UserAffiliations
-    {...defaultProps}
-    {...props}
-  />
+  <QueryClientProvider client={queryClient}>
+    <UserAffiliations
+      {...defaultProps}
+      {...props}
+    />
+  </QueryClientProvider>
 );
 
-describe('render UserAffiliations accordion component', () => {
+describe('UserAffiliations', () => {
   beforeEach(() => {
+    useUserAffiliationsMutation.mockClear().mockReturnValue({ handleAssignment: () => [], isLoading: false });
     useUserAffiliations
       .mockClear()
-      .mockReturnValue({ affiliations, totalRecords: affiliations.length });
+      .mockReturnValue({ affiliations, totalRecords: affiliations.length, isLoading: false, handleAssignment: () => [{}], refetch: () => {} });
   });
 
   it('should render a list of user affiliations', () => {
@@ -46,7 +67,38 @@ describe('render UserAffiliations accordion component', () => {
     renderPatronBlock();
 
     const primaryTenantListItem = screen.getByText(affiliations.find(({ isPrimary }) => isPrimary).tenantName);
-
     expect(primaryTenantListItem).toHaveClass('primary');
+  });
+
+  it.each`
+    status
+    ${'error'}
+    ${'success'}
+  `('should show $status message on click saveAndClose button', async ({ status }) => {
+    let mockErrorData = [];
+    if (status === 'error') {
+      mockErrorData = [{
+        'message': 'User with id [0c50701e-45ff-4a2e-bff0-11bd5610378d] has primary affiliation with tenant [mobius]. Primary Affiliation cannot be deleted',
+        'type': '-1',
+        'code': 'HAS_PRIMARY_AFFILIATION_ERROR'
+      },
+      {
+        'message': 'Some error message',
+        'type': '-1',
+        'code': 'GENERIC_ERROR'
+      }];
+    }
+
+    getResponseErrors.mockClear().mockReturnValue(mockErrorData);
+    renderPatronBlock();
+
+    const assignButton = screen.getByText('ui-users.affiliations.section.action.edit');
+    userEvent.click(assignButton);
+    const listOfAssignedTenants = await screen.findAllByRole('checkbox');
+    expect(listOfAssignedTenants).toHaveLength(2);
+    const saveAndCloseButton = screen.getByText('ui-users.saveAndClose');
+    userEvent.click(saveAndCloseButton);
+    await waitForElementToBeRemoved(() => screen.queryByText('ui-users.affiliations.manager.modal.title'));
+    expect(screen.queryByText('ui-users.saveAndClos')).toBeNull();
   });
 });